rpi-recognition/requirements.txt
pillow

-i https://download.pytorch.org/whl/cpu

torch 
torchvision
torchaudio
